Refactor client tests in multiple files
